Reorganize files into a more flat hierarchy
